requirements.txt
Django==5.1.2
dj-database-url
boto3
whitenoise[brotli]
easy-thumbnails
daphne
gunicorn
uvicorn[standard]
Twisted[tls,http2]
websockets
wsproto
argon2-cffi
psycopg[binary]
unpoly
stripe
Pillow
geoip2
python-magic
certifi 

django-storages
django-anymail
django-countries
django-debug-toolbar
django-filter
django-htmx
django-minify-html
django-browser-reload
django-forms-dynamic
django-template-partials
django-trix-editor
